Update license headers for Dgraph Community License files.
* Rename backup_oss.go -> backup.go
* Rename backup.go -> backup_ee.go
